remove HTTP layer from persistence, mount DB directly in manifesto model, UI talks to manifesto model facade instead now
