+++ services/docker-compose.yml
@@ -2,12 +2,12 @@ version: '3'
 services:
   persistence:
     build: persistence
-    ports:
-     - "0.0.0.0:8888:8888"
+#    ports:
+#     - "0.0.0.0:8888:8888"
     environment:
-      - PYTHONUNBUFFERED=0 # unbuffered I/O with print etc. to always print to the console
-      - HTTP_PORT=8888
-      - MANIFESTO_MODEL_HTTP_PORT=5000
+#      - PYTHONUNBUFFERED=0 # unbuffered I/O with print etc. to always print to the console
+#      - HTTP_PORT=8888
+#      - MANIFESTO_MODEL_HTTP_PORT=5000
       - DB_PATH=/db/active-manifesto.db
     volumes:
       # mount host database inside container
@@ -22,7 +22,8 @@ services:
     environment:
       - PYTHONUNBUFFERED=0 # unbuffered I/O with print etc. to always print to the console
       - HTTP_PORT=8080
-      - PERSISTENCE_HTTP_PORT=8888
+      - MANIFESTO_MODEL_HTTP_PORT=5000
+#      - PERSISTENCE_HTTP_PORT=8888
     volumes:
       # mount development files to app
       - ./user_interface:/app
@@ -34,7 +35,10 @@ services:
     environment:
       - PYTHONUNBUFFERED=0 # unbuffered I/O with print etc. to always print to the console
       - HTTP_PORT=5000
-      - PERSISTENCE_HTTP_PORT=8888
+      - DB_PATH=/db/active-manifesto.db
+#      - PERSISTENCE_HTTP_PORT=8888
     volumes:
+      # mount host database inside container
+      - ./db/active-manifesto.db:/db/active-manifesto.db
       # mount development files to app
       - ./manifesto_model:/app
