+++ services/manifesto_model/api.py
@@ -8,53 +8,20 @@ import requests
 from apscheduler.schedulers.background import BackgroundScheduler
 from flask import Flask, request, jsonify
 
+from sqlite_wrapper import get_texts_with_labels, insert_into, get_texts_only, get_texts_with_ids
+
 from classifier import Classifier
 
 app = Flask(__name__)
 
-DEBUG = os.environ.get('DEBUG') is not None
+DEBUG = True  # os.environ.get('DEBUG') is not None
 VERSION = 0.1
-
-PERSISTENCE_HTTP_PORT = int(os.environ.get('PERSISTENCE_HTTP_PORT'))
 
 classifier = Classifier()
 
-
-def retrain():
-    """
-    Fetches latest training data from persistence service and retrains the manifesto model with it.
-    """
-    print('retrain called')
-
-    url = 'http://persistence:{}/training_texts'.format(PERSISTENCE_HTTP_PORT)
-    print('requesting training data from persistence...')
-    r = requests.get(url=url)
-    print(r.status_code)
-
-    response_data = r.json()['data']
-    texts = list(map(lambda entry: entry['text'], response_data))
-    labels = list(map(lambda entry: entry['label'], response_data))
-
-    print('training on', len(texts), 'samples')
-    classifier.train(texts, labels)
+n_all_samples = 1000000
 
 
-# give some time for persistence to be up
-time.sleep(5)
-retrain()
-
-scheduler = BackgroundScheduler()
-scheduler.add_job(retrain, 'interval', minutes=60)
-scheduler.start()
-
-
-@app.route("/predict", methods=['POST'])
-def predict():
-    text = request.form['text']
-    return jsonify(classifier.predict(text))
-
-
-@app.route("/estimate_uncertainty", methods=['POST'])
 def estimate_uncertainty():
     """
     estimates text uncertainties.
@@ -88,6 +55,76 @@ def estimate_uncertainty():
     return jsonify({"data": response_data})
 
 
+def retrain():
+    """
+    Fetches latest training data from persistence service and retrains the manifesto model with it.
+    """
+    response_data = get_texts_with_labels(n_all_samples)
+    texts = list(map(lambda entry: entry['statement'], response_data))
+    labels = list(map(lambda entry: entry['label'], response_data))
+    print('training on', len(texts), 'samples')
+    classifier.train(texts, labels)
+
+
+scheduler = BackgroundScheduler()
+scheduler.add_job(retrain, 'interval', minutes=60)
+scheduler.start()
+
+
+@app.route("/texts_and_labels", methods=['POST'])
+def texts_and_labels():
+    """
+    Stores labels for the specified text ids.
+
+    expects a POST-body in the format:
+    {
+        "data": [
+            {"text_id": 17342, "label": "left"},
+            {"text_id": 873, "label": "neutral"},
+            ...
+        ]
+    }
+    """
+    texts_with_labels = json.loads(request.get_data(as_text=True))['data']
+    text_ids = map(lambda entry: entry['text_id'], texts_with_labels)
+    labels = map(lambda entry: entry['label'], texts_with_labels)
+
+    insert_into(text_ids, labels, 'user')
+    n_inserts = len(texts_with_labels)
+
+    return jsonify({'n_inserted': n_inserts}), 201
+
+
+@app.route("/prioritized_texts", methods=['GET'])
+def prioritized_texts():
+    """
+    Retrieves all of the political statements from the database, prioritizes them with the manifesto model
+    and returns at most as many texts given in the GET parameter `n`, ordered by their uncertainty.
+
+    :return: {
+        'data': [
+            {'text_id': 1, 'label': 'left'},
+            ...
+        ]
+    }
+    """
+    print("getting prioritized texts")
+    n_texts = int(request.args.get('n'))
+    texts = get_texts_only(n_all_samples)
+    prios_texts = classifier.prioritize(map(lambda t: t['statement'], texts))
+    # print(prios_texts)
+    texts_priotized = np.array(texts)[np.array(prios_texts)].tolist()
+    # print(texts_priotized)
+    return jsonify({'data': texts_priotized[:n_texts]})
+
+
+@app.route("/predict", methods=['POST'])
+def predict():
+    text = request.form['text']
+    return jsonify(classifier.predict(text))
+
+
 if __name__ == "__main__":
+    retrain()
     port = int(os.environ.get('HTTP_PORT'))
     app.run(host='0.0.0.0', port=port, debug=DEBUG)
