+++ services/manifesto_model/sqlite_wrapper.py
