+++ services/user_interface/main.py
@@ -11,7 +11,7 @@ app = Flask(__name__)
 app.debug = True
 app.secret_key = 'development'
 
-PERSISTENCE_HTTP_PORT = os.environ.get('PERSISTENCE_HTTP_PORT')
+MANIFESTO_MODEL_HTTP_PORT = os.environ.get('MANIFESTO_MODEL_HTTP_PORT')
 HTTP_PORT = int(os.environ.get('HTTP_PORT'))
 
 
@@ -22,7 +22,7 @@ def user_labels():
     """
     labels = json.loads(request.get_data(as_text=True))
     # registered hostname of service in docker-compose network
-    url = 'http://persistence:{}/texts_and_labels'.format(PERSISTENCE_HTTP_PORT)
+    url = 'http://manifesto_model:{}/texts_and_labels'.format(MANIFESTO_MODEL_HTTP_PORT)
     print(labels)
     r = requests.post(
         url=url,
@@ -37,8 +37,8 @@ def user_labels():
 def get_samples():
     n_texts = request.args.get('n')
     r = requests.get(
-        'http://persistence:{}/prioritized_texts?n={}'.format(
-            PERSISTENCE_HTTP_PORT,
+        'http://manifesto_model:{}/prioritized_texts?n={}'.format(
+            MANIFESTO_MODEL_HTTP_PORT,
             n_texts
         )
     )
@@ -49,7 +49,7 @@ def get_samples():
 def swipe():
     return render_template(
         'swipe.html',
-        persistence_host='http://0.0.0.0:'+str(PERSISTENCE_HTTP_PORT)
+        # persistence_host='http://0.0.0.0:'+str(MANIFESTO_MODEL_HTTP_PORT)
     )
 
 
@@ -57,7 +57,7 @@ def swipe():
 def index():
     return render_template(
         'index.html',
-        persistence_host='http://0.0.0.0:'+str(PERSISTENCE_HTTP_PORT)
+        # persistence_host='http://0.0.0.0:'+str(MANIFESTO_MODEL_HTTP_PORT)
     )
